+++ NEWS.md
@@ -1,5 +1,6 @@
 # hlatools 0.0.6.9000
 
+* Add `exon()`, `intron()`, and `utr()` to extract feature sequences
 * Partial matching of allele names when subsetting `HLAAllele` objects.
 * `HLAAllele` objects include _locusname_.
 * Added a `NEWS.md` file to track changes to the package.